requirements.txt
Click==7.0
docker==4.1.0
PyYAML==5.3
flake8
